moved ft.del  to doc_del
